docker/requirements.txt
# Spark 관련 패키지
pyspark==3.5.0  # Spark Python API

# AWS S3 연동
boto3==1.26.116  # AWS SDK for Python

# .env 파일 로드
python-dotenv==1.0.0  # 환경 변수 관리

pandas==1.5.3  # 데이터 분석 라이브러리
numpy==1.24.2  # 수치 계산 라이브러리

tqdm==4.64.1  # 진행률 표시 라이브러리

apache-airflow-providers-apache-spark
apache-airflow-providers-openlineage>=1.8.0